+++ src/hooks.ts
@@ -23,7 +23,12 @@ export const HOOKS: { [k: string]: any[] } = {
     createRxDatabase: [],
     preCreateRxCollection: [],
     createRxCollection: [],
-    /**
+     /**
+     * runs at the end of the destroy-process of a collection
+     * @async
+     */
+    postDestroyRxCollection: [],
+   /**
      * functions that get the json-schema as input
      * to do additionally checks/manipulation
      */